+++ config/test.exs
@@ -1,16 +1,15 @@
 use Mix.Config
 
-config :faktory_worker_ex,
+config :faktory_worker_ex, Test.Client,
   port: 7421,
-  client: [pool: 5],
-  use_tls: false,
-  workers: [
-    default: [
-      concurrency: 2
-    ],
-    middleware: [
-      concurrency: 2,
-      middleware: [BadMath],
-      queues: ["middleware"]
-    ]
-  ]
+  pool: 5
+
+config :faktory_worker_ex, Test.DefaultWorker,
+  port: 7421,
+  concurrency: 2
+
+config :faktory_worker_ex, Test.MiddlewareWorker,
+  port: 7421,
+  concurrency: 2,
+  middleware: [BadMath],
+  queues: ["middleware"]