+++ lib/faktory/application.ex
@@ -3,7 +3,6 @@ defmodule Faktory.Application do
   use Application
 
   def start(_type, _args) do
-    Faktory.Configuration.init
     Faktory.Supervisor.start_link
   end
 end