+++ lib/faktory/supervisor/clients.ex
@@ -2,25 +2,30 @@ defmodule Faktory.Supervisor.Clients do
   @moduledoc false
   use Supervisor
 
-  def start_link(config) do
-    Supervisor.start_link(__MODULE__, config, name: __MODULE__)
+  def start_link(config_modules) do
+    Supervisor.start_link(__MODULE__, config_modules, name: __MODULE__)
   end
 
-  def init(config) do
-    Supervisor.init(children(config), strategy: :one_for_one)
+  def init(config_modules) do
+    children(config_modules) |>
+      Supervisor.init(strategy: :one_for_one)
   end
 
-  def children(nil), do: []
+  def children([]), do: []
 
-  def children(config) do
-    pool_options = [
-      name: {:local, config.name},
-      worker_module: Faktory.Connection,
-      size: config.pool,
-      max_overflow: 2
-    ]
+  def children(config_modules) do
+    Enum.map(config_modules, fn module ->
+      config = module.config
 
-    [:poolboy.child_spec(config.name, pool_options, config)]
+      pool_options = [
+        name: {:local, module},
+        worker_module: Faktory.Connection,
+        size: config.pool,
+        max_overflow: 2
+      ]
+
+      :poolboy.child_spec(module, pool_options, config)
+    end)
   end
 
 end