+++ lib/faktory/worker.ex
@@ -146,13 +146,13 @@ defmodule Faktory.Worker do
     ((now_in_ms() - t) / 1000) |> Float.round(3)
   end
 
-  defp with_conn(%{name: pool}, f) do
+  defp with_conn(%{module: pool} = state, f) do
     try do
       :poolboy.transaction(pool, f)
     catch
       :exit, {:timeout, _} ->
         Logger.warn("connection pool timeout")
-        with_conn(%{config_module: pool}, f)
+        with_conn(state, f)
     end
   end