+++ test/support/bad_math.ex
@@ -1,12 +1,16 @@
 defmodule BadMath do
   use Faktory.Middleware
 
-  def call(job, chain, f) do
+  def call(job, chain, f, options \\ nil) do
     job = Map.update! job, "args", fn [pid, x, y | []] ->
       [pid, x+1, y+1]
     end
 
-    f.(job, chain)
+    if options do
+      f.(job, chain, options)
+    else
+      f.(job, chain)
+    end
   end
 
 end